requirements.txt
aiofiles==24.1.0
anyio==4.4.0
appdirs==1.4.4
asgi-csrf==0.9
asgiref==3.8.1
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-default-group==1.2.4
cryptography==42.0.8
datasette==0.64.6
datasette-cluster-map==0.18.2
datasette-embeddings==0.1a3
datasette-enrichments==0.4.2
datasette-leaflet==0.2.2
datasette-secrets==0.2
flexcache==0.3
flexparser==0.3.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
hupper==1.12.1
idna==3.7
itsdangerous==2.2.0
janus==1.0.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mergedeep==1.3.4
numpy==2.0.0
pandas==2.2.2
Pint==0.24.1
pluggy==1.5.0
pycparser==2.22
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
setuptools==70.1.1
six==1.16.0
sniffio==1.3.1
sqlite-fts4==1.0.3
sqlite-utils==3.36
tabulate==0.9.0
tqdm==4.66.4
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
uvicorn==0.30.1
WTForms==3.1.2
